Create Foo_textdisplay
Basic script to format the output from Foobar2K text component display module (foo_textdisplay).  Output will show you relevant information to playing file regardless of format, in regards to format (codec/profile) details for that particular format will be displayed conditionally to maintain format and save char space in small window implementations.  Includes a progress bar using UTF boxes to indicate player position and relative remaining play time of current media that is PLAYING, not selected.  Focus is always on playing media, not on selection.  

Timers are shown for current media as well (Remaining, Position(current), Total Duration).


This is my first beta release of the scripting, it works well so far, but I'm sure I will find some things that need touching up.  
For example: I will fix current code so that when MultiCH sources are played, in other words, if the return result on %channels% is other than '2' or 'stereo' or 'mono', indicating a surround source of 3,4,5,6,7,8-16 channels is in current playback, it will display "#CH - Multi-Channel Surround" where # = number of playback channels.  For example currently this script will display a DTS re-encoded FLAC file as 6CH, instead of giving a little more detail.  HOWEVER, this can be misleading as there is no MULTICH decoding occurring in most cases when reading a FLAC/MP3/PCM source,  as the files will general be time-synced & CH driven within the encoding process itself and output is specified within that format for each channel within the finalized format, for instance with FLAC it's really just playing back the data indicated for each channel along the timeline of the whole file, no other high-level decoding occurs, so that said... if indicator is changed to show "Multi-Channel Surround" or something of the sort. I need to see if I can make it clear that it is not decoding a MCH source like DTS/+, Dolby Digital/+, Dolby Atmos, Etc.  If this is the case I will need to take output from the DSP chain to verify how the sound is being processed and relay that into the text window.