Update Foo_textdisplay_Updated
Latest iteration, I don't know if I would call this cleaned up.  I have to reparse it by hand again, I'm getting strange errors when trying to brighten/dim text, there might be an errant empty space somewhere or something.  Although after playing around with this for a couple hours, it might be a bug in the beta of the component itself, in which case I will need to see if I can get into the source code on the project.  I don't think it is GIT.

1. Added a string search to insert a CRLF after the first '(' found in %title%, this will separate out the parenthesis content from the first section of the title, allowing better format flow instead of a long running line.
2. Adjusted line positions and colors (where possible, new bug may have been found in source component causing formatting problems)
3. Added pads where needed.
4. added vertical pipe/separator '|' to various subsections to improve readability of output.
5. Changed channels function, will now display nCH - In <mode> Using <MP3 MODE>  , for example:
a.   2CH - In Stereo - Using Joint Stereo
b . 2CH - In Stereo
c. 6CH - In 6CH  (I need to create a filtered statement to insert the word "Multi-CH' or 'Surround'... ToDO LIST!